Answer with the width the dataset fits, and guard it where data is built
ZONE_ID_RESULT_DTYPE was int32 because ``zone_ids`` is stored as uint16,
"whose upper half int16 could not hold". That reasons from the container's
range rather than from the values: a zone id is an index into the zone
names, ~450 of them against int16's 32,767, and the shortcut table already
stores those same ids as int16 today. It is a width chosen for headroom,
which is the case the narrow-dtype rule names - and it doubled the primary
batch return array, 40 MB rather than 20 MB for ten million points, in a
library whose stated constraint is constrained-memory containers.

Narrowed to int16, with the guard that makes a width chosen by fit safe.
``validate_shortcut_index`` now refuses a data directory whose zone count
outgrows the id width, naming the count, the ceiling, both constants to
widen and the version bumps that follow - and it already runs in both
places the rule asks for: over what ``scripts/file_converter.py`` just
wrote, and over the committed data in the test suite. Nothing new runs
when a finder is constructed.

The constant moves to configs.py: it is a data-derived width, the converter
side now reads it too, and declaring it once is what keeps the two from
disagreeing about how wide an id is.

+++ timezonefinder/global_functions.py
@@ -115,13 +115,13 @@ def timezone_ids_at(
     Equivalent to :meth:`TimezoneFinder.timezone_ids_at`, which documents the arguments,
     the ``on_invalid`` policies and every error raised.
 
-    :return: one ``int32`` timezone id per input coordinate, or ``NO_ZONE_ID`` (``-1``)
+    :return: one ``int16`` timezone id per input coordinate, or ``NO_ZONE_ID`` (``-1``)
         where the scalar lookup would answer ``None``
 
     Example:
         >>> ids = timezone_ids_at(lngs=[13.358, 2.3522], lats=[52.5061, 48.8566])
         >>> ids.dtype
-        dtype('int32')
+        dtype('int16')
     """
     return _get_tf_instance().timezone_ids_at(
         lngs=lngs, lats=lats, on_invalid=on_invalid
